feat: move MatchingStatus into new file `schema.py`
