+++ perception_eval/perception_eval/common/dataset_utils.py
@@ -34,8 +34,8 @@ from perception_eval.common.label import LabelType
 from perception_eval.common.label import TrafficLightLabel
 from perception_eval.common.object2d import DynamicObject2D
 from perception_eval.common.object import DynamicObject
-from perception_eval.common.status import FrameID
-from perception_eval.common.status import Visibility
+from perception_eval.common.schema import FrameID
+from perception_eval.common.schema import Visibility
 from PIL import Image
 from pyquaternion.quaternion import Quaternion
