+++ perception_eval/perception_eval/common/object2d.py
@@ -20,8 +20,8 @@ from typing import Tuple
 
 import numpy as np
 from perception_eval.common.label import Label
-from perception_eval.common.status import FrameID
-from perception_eval.common.status import Visibility
+from perception_eval.common.schema import FrameID
+from perception_eval.common.schema import Visibility
 from shapely.geometry import Polygon
