+++ perception_eval/perception_eval/visualization/perception_visualizer3d.py
@@ -30,7 +30,7 @@ from matplotlib.transforms import Affine2D
 import numpy as np
 from perception_eval.common.evaluation_task import EvaluationTask
 from perception_eval.common.object import DynamicObject
-from perception_eval.common.status import FrameID
+from perception_eval.common.schema import FrameID
 from perception_eval.config import PerceptionEvaluationConfig
 from perception_eval.evaluation import DynamicObjectWithPerceptionResult
 from perception_eval.evaluation import PerceptionFrameResult
