+++ perception_eval/test/util/dummy_object.py
@@ -20,7 +20,7 @@ from perception_eval.common.label import AutowareLabel
 from perception_eval.common.label import Label
 from perception_eval.common.object2d import DynamicObject2D
 from perception_eval.common.object import DynamicObject
-from perception_eval.common.status import FrameID
+from perception_eval.common.schema import FrameID
 from pyquaternion.quaternion import Quaternion
